hw2: migrate display update to the wdt timer
The assignment requires that the wdt is enabled in timer mode, but doesn't
actually specify the role of the wdt. With this in mind, we appropriate the wdt
timer to driver the display update. This is done because wdt interval
adjustment is trivial and understood, whereas ttc interval update is currently
not.

In an ideal world, we'd drive the display update on the ttc context, and adjust
the ttc interval dynamically. Since I don't know how to do this and I'm running
out of time ...